Applied scripts/license.sh.

incremental_compaction_test adjusted for update_compaction_history.

* 'master' of github.com:scylladb/scylla: (97 commits)
  raft: re-advertise gossiper features when raft feature support changes
  raft: move tracking `SUPPORTS_RAFT_CLUSTER_MANAGEMENT` feature to raft
  gms: feature_service: update `system.local#supported_features` when feature support changes
  test: cql_test_env: enable features in a `seastar::thread`
  scripts: Detect remote branch to fetch submodules from
  test/boost: mutation_reader_test: add tests for generating reader
  test: export squash_mutations() into lib/mutation_source_test.hh
  readers: add next partition adaptor
  readers: implement generating_reader from v1 generator via adaptor
  readers: upgrade_to_v2(): reimplement in terms of upgrading_consumer
  readers: add upgrading_consumer
  readers: generating_reader: use noncopyable_function<>
  readers: merge generating.hh into generating_v2.hh
  readers/generating.hh: return v2 reader from make_generating_reader()
  raft: make raft requests abortable
  replica: distributed_database: populate_column_family: trigger offstrategy compaction only for the base directory
  test/boost/mutation_reader_test: remove unused puppet_reader
  test/boost/flat_mutation_reader_test: test_flat_mutation_reader_consume_single_partition: make ckrange conditionally inclusive
  system_keyspace: Make load_host_ids non-static
  system_keyspace: Make load_tokens non-static
  ...

+++ message/messaging_service.cc
@@ -90,7 +90,7 @@
 #include <boost/range/adaptor/filtered.hpp>
 #include <boost/range/adaptor/indirected.hpp>
 #include "frozen_mutation.hh"
-#include "flat_mutation_reader.hh"
+#include "readers/flat_mutation_reader.hh"
 #include "streaming/stream_manager.hh"
 #include "streaming/stream_mutation_fragments_cmd.hh"
 #include "locator/snitch_base.hh"
@@ -415,7 +415,7 @@ future<> messaging_service::stop() {
             for (auto verb = messaging_verb::MUTATION; verb < messaging_verb::LAST;
                     verb = messaging_verb(int(verb) + 1)) {
                 if (_rpc->has_handler(verb)) {
-                    mlogger.error(" - {}", verb);
+                    mlogger.error(" - {}", static_cast<int>(verb));
                 }
             }
